requirements.txt
labelbox==3.65.0
Pillow==10.2.0
